// src/app/models/enums/section-keys.enum.ts
export enum UISectionKeysEnum {
    ADMIN_PROFILE = 'ADMIN_PROFILE',
    ADMIN_SERVICES = 'ADMIN_SERVICES',
    ADMIN_CALENDAR = 'ADMIN_CALENDAR',
    ADMIN_RESERVATIONS = 'ADMIN_RESERVATIONS',
    ADMIN_HOLIDAYS = 'ADMIN_HOLIDAYS',
    PUBLIC_PROFILE = 'PUBLIC_PROFILE',
    PUBLIC_RESERVATIONS = 'PUBLIC_RESERVATIONS',
    PUBLIC_SERVICES = 'PUBLIC_SERVICES',
    PUBLIC_BUSINESSES = 'PUBLIC_BUSINESSES',
}